refactor(abstract-utxo): drop DescriptorMap/SignerKey/sumValues from wasmUtil
Move the DescriptorMap type alias to descriptor/index.ts (the natural
location), inline SignerKey at its single call site, and inline the
sumValues one-liner at the two explainPsbt/parse call sites.

Refs: T1-3279
